add conformance test, rename to parse_record
