+++ rust/Cargo.toml
@@ -51,7 +51,7 @@ serde = { version = "1.0.145", features = ["derive"] }
 serde_json = "1"
 simplelog = "0.12"
 tempfile = "3.3"
-tokio = { version = "1", features = ["io-util", "macros", "rt"] }
+tokio = { version = "1", features = ["io-util", "macros", "rt", "fs"] }
 
 [[bench]]
 name = "reader"
