+++ rust/examples/conformance_reader_async.rs
@@ -1,0 +1,37 @@
+#[path = "common/serialization.rs"]
+mod serialization;
+
+use serde_json::{json, Value};
+
+use serialization::as_json;
+use std::env;
+use std::process;
+use tokio::fs::File;
+
+use tokio;
+
+#[tokio::main(flavor = "current_thread")]
+async fn main() {
+    let args: Vec<String> = env::args().collect();
+    if args.len() < 2 {
+        eprintln!("Please supply an MCAP file as argument");
+        process::exit(1);
+    }
+    let file = File::open(&args[1]).await.expect("couldn't open file");
+    let mut reader = mcap::tokio::RecordReader::new(file);
+
+    let mut json_records: Vec<Value> = vec![];
+    let mut buf: Vec<u8> = Vec::new();
+    while let Some(op) = reader
+        .next_record(&mut buf)
+        .await
+        .expect("failed to read next record")
+    {
+        if op != mcap::records::op::MESSAGE_INDEX {
+            let parsed = mcap::parse_record(op, &buf[..]).expect("failed to parse record");
+            json_records.push(as_json(&parsed));
+        }
+    }
+    let out = json!({ "records": json_records });
+    print!("{}", serde_json::to_string_pretty(&out).unwrap());
+}
